requirements.txt
# 必须通过pip安装的第三方库
pillow>=9.0.0  # 图像处理功能